unify "by user" reports
