+++ templates/reporting/report_by_user.html.twig
@@ -1,6 +1,6 @@
 {% extends 'reporting/layout.html.twig' %}
 
-{% block report_title %}{{ 'report_user_month'|trans({}, 'reporting') }}{% endblock %}
+{% block report_title %}{{ report_title|trans({}, 'reporting') }}{% endblock %}
 
 {% block report %}
 
@@ -14,7 +14,7 @@
     {% embed '@AdminLTE/Widgets/box-widget.html.twig' %}
         {% import "macros/widgets.html.twig" as widgets %}
         {% block box_before %}
-            {{ form_start(form, {'action': path('report_user_month'), 'attr': {'class': 'form-inline'}}) }}
+            {{ form_start(form, {'attr': {'class': 'form-inline'}}) }}
         {% endblock %}
         {% block box_after %}
             {{ form_end(form) }}
